+++ mathics/core/symbols.py
@@ -554,9 +554,6 @@ class Symbol(Atom, NumericOperators, EvalMixin):
                 1,
             ]
 
-    def is_true(self) -> bool:
-        return self is SymbolTrue
-
     def user_hash(self, update) -> None:
         update(b"System`Symbol>" + self.name.encode("utf8"))