+++ CHANGES.rst
@@ -61,7 +61,8 @@ Internals
 * ``N[_,_,Method->method]`` was reworked (Issue #137).
 * ``InstanceableBuiltin`` -> ``BuiltinElement``
 * ``BoxConstruction`` -> ``BoxExpression``
- 
+* the method ``Element.is_true()`` was removed in favor of `is SymbolTrue`
+
 
 Package update
 ++++++++++++++